requirements.txt
openai>=1.51.0
requests>=2.32.0
pypdf>=5.1.0
PyMuPDF>=1.24.0
tenacity>=9.0.0
PyYAML>=6.0.2
numpy>=2.1.2
sentence-transformers>=3.2.0
feedparser>=6.0.11
arxiv>=2.1.3
biopython>=1.84
python-dateutil>=2.9.0
openreview-py>=1.45.0
pycryptodome>=3.21.0
curl_cffi>=0.7.3
playwright>=1.52.0
vllm>=0.6.0; platform_system != "Windows"
